depthcharge: Use the new block_dev_init and block_dev_refresh.
This change removes manual calls to initialize and poll for storage devices in
favor of the new block_dev_init and block_dev_refresh calls. Also, storage
initialization and polling now happens in response to vboot asking about them.
The resulting behavior shouldn't change, but this ensures that that work is
only done if vboot is going to need it and in time for vboot to use it.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link.
BRANCH=None

Change-Id: Id8070ec6840999ea3a928929210949995abc5aa1
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/28966
Reviewed-by: Ronald G. Minnich <rminnich@google.com>