depthcharge: Generalize block device initialization and refresh.
To avoid hardcoding what types of devices to initialize and refresh, this
change adds a new list of data structures which represent storage controller
hardware. When a new type/instance of a storage controller is built in, it can
register an init function and add itself to the list of storage controllers.
Then later, when we need to look at storage, it can be called back into to do
whatever it needs to do.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link.
BRANCH=None

Change-Id: I9786869af346c0b4d92173a1f86a363835b1bf26
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/29123
Reviewed-by: Ronald G. Minnich <rminnich@google.com>