depthcharge: Initialize the EC driver only when necessary.
This change delays initializing the EC driver until vboot asks depthcharge to
do something EC related. We'll be sure to initialize things if and only if
we're going to actually use it, and as late as possible to avoid duplicating
work between the RO and RW firmwares.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link. Verified that the EC id string was retrieved
and printed correctly.
BRANCH=None

Change-Id: Ifc1c9d4539415227a9a1f578e1818ea6ef04de6a
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/29127
Reviewed-by: Bill Richardson <wfrichar@google.com>
Reviewed-by: Stefan Reinauer <reinauer@google.com>