depthcharge: Split out arch dependent trampoline code.
The trampoline code which switches stacks to the one in the trampoline is
architecture dependent because it uses a little bit of inline assembly. The
rest of the trampoline which loads and enters an ELF image is generic. When
architecture specific code is split out, the ELF loading code can be shared
while the other code will have to be implemented for each arch.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link. Forced an RW boot on Link.
BRANCH=None

Change-Id: I223d25645790e149d98680fb17c006e9bbceda52
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/29237
Reviewed-by: Stefan Reinauer <reinauer@google.com>