depthcharge: Rearrage how config information is stored for a given board.
Right now, configuration is tied to the ebuild used to build depthcharge
because it has to copy config.$(board} to .config and then run make oldconfig.
That's a little clumsy to use if you want to build a board manually since you
have to copy over the .config yourself, and it's fairly easy to accidentally
leave that hidden file lying around to interfere with later builds. You can
also forget to copy over config.${board} after modifying it and end up with a
stale config. Instead, we can take advantage of the defconfig mechanism
(we're really defining default configs anyway), creatively define where the
default config lives, and have make handle things for us in line with how
Kconfig normally works.

In addition to the Kconfig config file, we also have an fdt per board to
provide the fmap some of our tools require. Having two files per board
bouncing around in the same directory is a bit messy even with two boards and
makes it harder than it should be to see what sets of configs are available.
Also, having a generically named device tree may imply that we're actually
using it for something ourselves, or that it has more info in it than the
fmap. To solve these two problems, the device trees were renamed to fmap.dts,
and the config and fmap were put into directories named after the board they
describe.

The parent directory was renamed from config to board so that it more
specifically describes what's in it, information about a particular board.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link, with a corresponding ebuild change.
BRANCH=None

Change-Id: Id4980bcaa420886dba75a7bf7b6a7b127e107952
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/29243
Reviewed-by: Stefan Reinauer <reinauer@google.com>