depthcharge: Move the linker script into src/image.
Since the linker script has to do with building depthcharge, put it with the
other depthcharge source files. Because it determines where things go in the
final image, put it in the image directory. Having it where it is now isn't
terrible, but it's new home seems to make a little more sense.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link.
BRANCH=None

Change-Id: Icd515948f5f07ff9ea4379a34cef0d0e97975a9c
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/29244
Reviewed-by: Stefan Reinauer <reinauer@google.com>