depthcharge: Factor out the range data structure in the memory wipe code.
The memory wiping code uses a data structure to keep track of the ranges to
wipe which might be useful in other areas of depthcharge, for instance to keep
track of what areas of a software cache of the flash are valid and what parts
still need to be read in. This change pulls that data structure out of the
memory wipe code itself and makes it available generally.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link.
BRANCH=None

Change-Id: I5451941709f44ed44fe6dca84d8a070fb5afd2a6
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/29308
Reviewed-by: Stefan Reinauer <reinauer@google.com>