depthcharge: Initialize USB on demand.
The USB initialization code in libpayload actually enumerates the USB bus in
addition to finding the controllers in the first place. I didn't think this
was the case because there used to be an explicit call to usb_poll within the
USB initialization function. To avoid enumerating USB devices when we're not
actually going to use them, we put that off until someone tries to do
something related to USB. At present, that includes trying to read a character
of input (USB keyboards) or getting information about USB storage devices.

Because libpayload probably can't handle usb_initialize being called more than
once, depthcharge wraps it with a function which tracks whether or not it's
been called, and if so just returns immediately.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link.
BRANCH=None

Change-Id: Ida2884ae0f4774d417c62c3e0b2b7ca2e3d2f33c
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/29315
Reviewed-by: Stefan Reinauer <reinauer@google.com>