depthcharge: Tell vboot if it should do software sync, and handle EC reboots.
Vboot needs to be told if it should be worried about EC software sync. It will
tell us if we need to reboot the EC to get it back into RO, so we need to
handle that too.

BUG=chrome-os-partner:16678
TEST=Ran the firmware_RONormalBoot FAFT test and saw it update the EC RW and
successfully reboot back into RO during the cleanup part of the test.
BRANCH=None

Change-Id: I06680d6c0c4fca97b9884baa8f6e0bf0960bdbe0
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/30392
Reviewed-by: Duncan Laurie <dlaurie@google.com>