depthcharge: Move the ACPI update after VbSelectAndLoadKernel.
That way the info put into the ACPI tables is as up to date as possible.
Otherwise, the RO/RW state of the EC may not be recorded accurately..

BUG=chrome-os-partner:16678
TEST=Ran the firmware_RONormalBoot FAFT test. It failed before this change
even though the EC was changing between RO and RW mode correctly because the
state was recorded incorrectly in the ACPI tables. After this change the test
passed.
BRANCH=None

Change-Id: Ib8b2cc8f543cf8a4ba96aca92fa0f8b370220ef8
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/30393