depthcharge: ec: Make the MKBP driver wait for the EC when it jumps to RW.
Otherwise, when that transition happens, there's a window where the EC won't
respond to commands but the driver has given control back to depthcharge
proper. It may be that the lost commands seem to succeed because they see the
result of the last command that was actually delivered, even if the actual
returned data is junk.

BUG=chrome-os-partner:16678
TEST=Ran the faft_lv3 suite and saw the RONormalBoot test fail because
depthcharge would halt when it got junk data when determining whether the EC
was in RO or RW mode. With this change the test passes, and the EC console
shows depthcharge waiting until the EC responds to the "Hello" command before
sending it other subsequent commands. When debugging previous problems with
this test, serial output was turned on which changed the timing and avoided
sending commands during the jump to RW.
BRANCH=None

Change-Id: Ia074052dcf56c16c42f7be782405983d9947451d
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/30394