depthcharge: Always hide the actual "printf" command line.
printf is used to print the abbreviated form of steps in the Makefile. If the
printf command line itself is printed when verbose mode is on, the the short
versions are printed twice, effectively. There's no additional value from that
command line. This change makes the printfs themselves unconditionally hidden.

BUG=chrome-os-partner:8339
TEST=Built with and without V=1 and saw that the printf command lines were
hidden.
BRANCH=None

Change-Id: I9a0bbfd324436e0fe5a557a5c53a36ac30e141a9
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/30396