depthcharge: Fix a bug filling in an SG in the AHCI driver.
When the code in the AHCI driver was streamlined and cleaned up, misplaced
parenthesis introduced a bug where the flags_size member of the SG was
computed. This change fixes that bug.

BUG=chrome-os-partner:16883
BUG=chrome-os-partner:8339
TEST=Ran the firmware_CgptState FAFT test until it attempted to update the
GPT. Before this change, the AHCI write would fail, corrupting the disk and
preventing further boots. After this change, the test completed successfully.
BRANCH=None

Change-Id: I9831521bef8ab5be838757caba1e1eff048dc9b6
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/30768
Reviewed-by: Walter Murphy <wmurphy@google.com>
Reviewed-by: Vadim Bendebury <vbendeb@google.com>