depthcharge: ec: Fix a memory management bug in the EC driver code.
In VbExEcHashRW, the RW hash returned by the EC is put into an
ec_response_vboot_hash structure allocated on the stack. A pointer is then
set to point to a buffer in the structure, and the function returns. That
means there's a pointer to live data that something is going to use which is
in a deallocated stack frame that's going to get clobbered if the stack grows
enough again.

This change nominally fixes the problem by making that structure static
instead of allocating it on the stack. Unfortunately that means the old value
will get clobbered if the function is called again, but that's the lesser of
two evils, and in practice the function isn't used that way now or in the near
forseable future. Another alternative would be to allocate a buffer every time
the function was called and let it leak, but that seemed less than desirable
as well.

This fix will need to be ported to other instances of this code, likely in
Coreboot and U-Boot.

BUG=chrome-os-partner:17143
BUG=chrome-os-partner:8339
TEST=Built and booted an RW image of depthcharge on Link. Saw it update the
EC RW every boot, and that the hash was wrong, but only a few bytes of it. The
hash could be further corrupted by calling functions within the vboot code
that compared it against the expected value. Ran again with this fix and saw
it boot and recognize that the hash matched and not update the EC RW
unecessarily.
BRANCH=None

Change-Id: Id0292ba4258a67b3e7cc3c544b0ea615b97b34c6
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/30904
Reviewed-by: Duncan Laurie <dlaurie@google.com>
Reviewed-by: Stefan Reinauer <reinauer@google.com>