depthcharge: Make the assignment to LIBGCC in Makefile.inc expand immediately.
Because the expression which generates the value for the variable LIBGCC uses
the "shell" make function and because its value isn't expected to change, it's
better to find that value once instead of each time the variable is used. This
makes the build every so slightly more efficient.

BUG=chrome-os-partner:8339.
TEST=Built depthcharge for Link. During the config phase, saw a message about
lpgcc not being found appear only once instead of several times.
BRANCH=None

Change-Id: Ia4e241accf48e6ce26d18f87c1db42c0c03972e1
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/31432
Reviewed-by: Ronald G. Minnich <rminnich@google.com>