depthcharge: Seperate generic code from time.c and add get_raw_timer_Value.
The new function get_raw_timer_value should be implemented to read whatever
time source is appropriate. Currently there is an implementation for x86 which
reads the TSC. Since all the architecture/timer specific code is now in that
one function, split out the generic stuff and move it back out of the arch
directory.

BUG=chrome-os-partner:17340
TEST=Built for Daisy and saw errors about rdtsc.h not being found go away.
Built and booted on Link.
BRANCH=None

Change-Id: I40a0289c0fc5706cb233003093800b1902cb5935
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/31438
Reviewed-by: Ronald G. Minnich <rminnich@google.com>