depthcharge: Change how architecture provided includes work.
There are includes which are per architecture which provide definitions which
are used by code outside the architecture, currently only types.h. The way
includes were being handled was to add the architecture directory to the
include search path and then to mirror the main directory heirarchy in the
arch directory. Then, when you wanted to, for example, include types.h for the
current arch, you'd #include "base/types.h" and get the right one.

There are two problems with this. First it's not obvious in the arch directory
which include files are intended for use outside the architecture and which
are for internal consumption. Second, it's not clear from outside the arch
directory what files are actually where they seem to be, and which are being
magically transplanted from the arch directory.

To address this problem, I added a new directory within arch which is
specifically for includes that are for external use called "includes". Within
it is a single directory called "arch". If we end up having so many files that
we need some stronger organization we can add subdirectories, but hopefully
that won't be necessary. When something outside of arch needs something
defined per architecture, then it can, for instance, #include "arch/types.h"
which is obviously provided by the current architecture.

Unfortunately in the one case where this mechanism is currently used,
"arch/types.h" is already a header provided by libpayload which overrides the
depthcharge version. I get around that by renaming the local version to
dc_types.h.

BUG=chrome-os-partner:17340
TEST=Built for Link. Built for Daisy and saw equivalent error messages.
BRANCH=None

Change-Id: I03278b595594dbf488c9479e9d864f3b6b9ea333
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/31639
Reviewed-by: Ronald G. Minnich <rminnich@google.com>