depthcharge: Wrap setting an initial post code in an arch specific function.
The read only and read/write main functions start by outputing a post code to
show that depthcharge is running even if no other output works. Unfortunately
that's architecture specific and can't be left in otherwise architecture
agnostic code. This change wraps it in a new function called "sign_of_life"
which can be implemented differently for different architectures.

To minimize dependence on the stack, proper linking, etc., the function is
defined as static inline and put directly in the architecture specific
header file.

BUG=chrome-os-partner:17340
TEST=Built and booted on Link. Built for Snow and saw errors about outb in
main go away.
BRANCH=None

Change-Id: Id2be74090574306b22114dab1bfd894deabf98e5
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/31641
Reviewed-by: Ronald G. Minnich <rminnich@google.com>