depthcharge: Decentralize Kconfig.
If Kconfig were scaled up so that more things were configurable (which is the
current plan) then having everything in one file would get to be unmanagable.
This change splits out the existing Kconfig so that image related options are
in image, and vboot related options are in vboot. The only thing left in the
main Kconfig is which architecture to use. The base of the Kconfig tree was
also moved up to src to keep it close to subordinate Kconfigs.

The original main Kconfig no longer has anything at all in it and just sources
the new main Kconfig in the src directory. It's needed so that the config
programs (menuconfig, etc) can find it.

Now that image related options are in the image directory, it felt awkward to
have to maintain a separate list of the same options in the main Makefile.
The list was moved into the Makefile.inc in the image directory to keep
everything together, and a comment was added to both the Kconfig and
Makefile.inc directing people from one to the other if they need to make
changes.

BUG=chrome-os-partner:17340
TEST=Built and booted on Link. Built on Daisy and saw the same error messages
as before. Ran "make menuconfig" inside the chroot and verified that things
looked sane and usable.
BRANCH=None

Change-Id: I7ca6c618bcf7882fab9dcfa9aa6597ac19653fc7
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/31643