depthcharge: Add some functions needed by libgcc.
These are:

int raise(int sig)
The standard library function which raises a signal. Since we don't have
those, this prints a message saying what happened and then hangs. This
function seems to only be called by the division function when it detects
a division by zero. If that happens we want to know about it so we can fix it,
and it should never happen in the absence of bugs.

__aeabi_unwind_cpp_pr0
A function that takes relatively complicated arguments and is used to unwind
the stack in the event of a C++ exception. I think the pr0 specifies the
personality. We're not in C++ and not using C++ exceptions, so we can just
ignore these requests.

BUG=chrome-os-partner:17340
TEST=Built for Snow and saw more linker errors go away.
BRANCH=None

Change-Id: Ide76988d4b51962f3a37daf5525f980fce598de2
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/31646