depthcharge: If we aren't using software sync, don't build in calls to the EC.
If EC software sync is turned off, then we shouldn't be calling into the EC
driver. This change adds stub implementations of the EC software sync related
functions which print an error and hang, and compiles those in selectively.

This will make it possible to disable the MKBP driver on Lumpy.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link, built for Lumpy. Used #error directives to
verify the right files were being compiled in for each system.
BRANCH=None

Change-Id: I9c43752f3c83f48a7ddc65220b291b4332e21dac
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/31655