depthcharge: Add in code to allow talking to the EC over I2C.
This interface is turned on for Daisy/Snow. The I2C bus and address to use are
configured using Kconfig variables. The generic version of the
mkbp_check_version is provided in a separate file and is compiled in when
building either the I2C or SPI interfaces. The I2C interface is now
parenthetically described as untested instead of unimplemented in Kconfig, and
the pass through send_command function was eliminated from mkbp.c

The generic mkbp_bus_check_version function is implemented a bit differently
than it was in U-Boot, partially due to the fact that it's compiled seperately
and can't get direct access to the static ec_command function in the main
mkbp.c. Instead of doing it's own half call to the EC_CMD_HELLO command while
still really doing the whole thing to avoid tripping up the EC, it now uses
the mkbp_test function which does essentially the same thing but which doesn't
make checking the result optional.

A few functions were renamed from mkbp_lpc_* to mkbp_bus_ so that they could
be provided by either the LPC bus driver, or the I2C or SPI bus drivers as
necessary. Because it's presumed there will be only one EC and therefore only
one bus for it to communicate over, no additional level of indirection is
necessary.

BUG=chrome-os-partner:17340
TEST=Built for Daisy and saw some linker errors related to mkbp go away.
BRANCH=None

Change-Id: Ic16648442c6a8a3794021c6e1e16bdf0cada837e
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/31665