depthcharge: Guard reading the EC in RW flag with CONFIG_EC_SOFTWARE_SYNC.
If we don't have EC software sync, we shouldn't need to read the GPIO which
says whether the EC is in RW. By putting this call behind an if based on that
compile time config option, we tell the compiler it's ok to drop the call from
the compiled version of the code entirely, and that tells the linker it can
drop the GPIO driver if nobody else is using it. This will result in a smaller
image on systems that don't need EC softwre sync.

BUG=chrome-os-partner:17340
TEST=Built and booted on Link. Built for Lumpy and saw the image size decrease
a bit. Built for Daisy and saw consistent linker errors.
BRANCH=None

Change-Id: Ib188cf69ad58f0f38466a152db8c1b364ba66a7c
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/31666