depthcharge: Seperate out the generic PCH driver and the lynxpoint lp one.
These were bundled together and differentiated at run time because the config
mechansim in depthcharge hadn't been fleshed out yet. Now that it has, there's
not much reason for these largely independent drivers to be squished together.
The pch_gpiobase function is now duplicated between them, but it's pretty
short so that isn't too bad. It could be split out and put in a common source
file, but for now it's not worth the complication.

BUG=chrome-os-partner:17340
TEST=Built and booted on Link. Built for Haswell and Lumpy. Built for Daisy
and saw consistent error messages.
BRANCH=None

Change-Id: I2970c8af5065314ea840800f6253cfbcb4f051b4
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/31668