depthcharge: Stub out using a disk for nvstorage, and make it configurable.
Systems without CMOS, aka everything but x86, need somewhere else to store the
small blob of non-volatile state vboot uses. This change stubs out some
functions which will ultimately manage that storage on a fixed disk. An option
was added to Kconfig which selects one mechanism or the other.

BUG=chrome-os-partner:17340
BUG=chrome-os-partner:16682
TEST=Built and booted on Link. Built for daisy and saw the build finish
successfully. The emerge did not, however, do to a missing fmap.dts.
BRANCH=None

Change-Id: I44dac37a047d0baa3a524572626c1e1612368117
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/31670