depthcharge: Surface the dummy power driver in Kconfig.
Instead of making the decision whether to use it the Makefile based on what
the arch is, push it into Kconfig and then just do what Kconfig says. This
makes it more consistent with other similar options elsewhere.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link. Built for Lumpy, Daisy, and Haswell. Verified
that the right driver was selected during the build for each.
BRANCH=None

Change-Id: I66b0ce7752dfe7978c86fd37104756801b4d1475
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/31902
Reviewed-by: Ronald G. Minnich <rminnich@google.com>