depthcharge: Clean up the default board configs.
There's no need to explicitly specify options which agree with the defaults.
Remove a header from when the configs were originally generated which is very
out of date. Put comments in that delimit different areas of the config, and
put things in sections into alphebetical order.

BUG=chrome-os-partner:8339
TEST=Built for Link, Haswell, Daisy, and Lumpy. Verified that the right thing
was built into each.
BRANCH=None

Change-Id: I7fc00b4545487d38a54c8b9ebee6bc8e1070a8f7
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/31901
Reviewed-by: Ronald G. Minnich <rminnich@google.com>