depthcharge: Make the Kconfig infrastructure less forgiving.
Because a lot of configuration is done by hand in depthcharge as opposed to
letting a tool do things, it's important that mistakes are caught early and
that things are actually configured the way they were intended to be. To help
assure that, this change makes some problems that would have just caused
warnings cause configuration to fail.

BUG=chrome-os-partner:8339
TEST=Built for Link, Haswell, Daisy and Lumpy. Inserted a line with garbage,
an unrecognized config option, an option which wasn't set to anything, and an
option set to something invalid one at a time in the Link default config and
saw that there was an appropriate error and that the build stopped.
BRANCH=None

Change-Id: Ide6dca48e68d2d5d452bf68ba344bb1ca2e4c428
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/31903
Reviewed-by: Ronald G. Minnich <rminnich@google.com>