depthcharge: Make the directory structure in drivers more consistent.
Most of the files in the drivers directory were in appropriate subdirectories,
but some where directly in the root. This was a bit inconsistent, and was
making the drivers directory feel pretty cluttered. To fix that I added
directories called tpm, video, and bus. I put the existing i2c directory and
the loose usb bus (as opposed to device) files into bus. The usb stack is
still in libpayload, those few files just wrap it so it can have on demand
initialization.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link. Built for Daisy.
BRANCH=None

Change-Id: Ie0417b9f8f6d01752520fb89600b335fcbda8c37
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/31942
Reviewed-by: Stefan Reinauer <reinauer@google.com>