depthcharge: Implement a real version of VbExEcGetExpectedRWHash.
The EC binary is moved into its own section of the FMAP so that it isn't
included in the verified RW area, and it's hash is put in its place. The EC
section is prepended with an index even though there's only one thing in it so
that we can tell how big the EC image is if it doesn't occupy its whole region
of the FMAP.

A comment in the code in "load_firmware" has also been updated to describe
what's currently in the verified area.

With this change in place, the RW penalty for depthcharge is a little less
than 40ms.

BUG=chrome-os-partner:17606
TEST=Built and booted on Link using an RW boot. Verified that the hash it
found in the image was correct, and that vboot accepted it and didn't update
the EC. Verified that performance was increased in line with earlier
experiments.
BRANCH=None

Change-Id: Iee83420c0a206e67224a13ca5953c0e96648fbaa
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/32107
Reviewed-by: Ronald G. Minnich <rminnich@google.com>
Reviewed-by: Randall Spangler <rspangler@google.com>