depthcharge: Create a "driver" interface for accessing flash contents.
On our x86 systems flash is memory mapped which makes it convenient to get to.
Because it's so convenient, there hasn't been a need for a more formal
interface to get at data stored there, you just had to compute a pointer and
grab what you wanted.

To support systems without memory mapped flash, and to facilitate better
performance if using the programmer interface directly is faster, this change
creates an interface which abstracts out exactly how flash is read or where
the contents are stored.

The new interface supports a single function, flash_read, which takes the
offset into the flash and a size and returns a pointer to the requested data
in memory. If the flash is memory mapped, the "driver" is just computing the
pointer which points to the data you want and tells you to access it in
place. If the driver actually goes out and reads the flash, it should put the
data into a buffer which will act as a cache and return a pointer to that.

A memory mapped flash driver and a dummy driver are provided as well.

BUG=chrome-os-partner:16683
TEST=With this and other changes, moved over to using this interface to access
flash. Built and booted on Link, built for Daisy, Lumpy and Haswell.
BRANCH=None

Change-Id: I455ea6edca3987b38fbb9580286158e8272f5391
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/32350
Reviewed-by: Ronald G. Minnich <rminnich@google.com>