depthcharge: Stop passing around a pointer to the main FMAP.
Needing a pointer to the fmap creates a chicken and egg problem as far as on
demand initialization since you have to know where the FMAP is before it may
have been found or processed. We only ever use one FMAP anyway, and passing
around a pointer which is always the same isn't very useful.

Also, since fmap_check_signature takes no arguments and is called
automatically on the only FMAP that matters, there's no need for it to be
available outside of fmap.c.

BUG=chrome-os-partner:16683
TEST=Built and booted on Link.
BRANCH=None

Change-Id: Ic4e253e1633fda8d716037b93452a8f38666c6cb
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/32352
Reviewed-by: Ronald G. Minnich <rminnich@google.com>