depthcharge: Ensure that the .cparams and .shared_data sections are not gc-ed.
These sections are intended to preserve space for data shared between the RO
and RW versions of depthcharge. Because they're not always referenced by both,
though, they may be dropped, and the things after them moved up to close the
gap. That means that if something after them *is* being shared between RO and
RW and actually being used, then it won't line up and you'll get the wrong
data.

BUG=chrome-os-partner:16683
TEST=Built without this change, used readelf to verify that the .shared_data
section was being garbage collected and later sections were moved up. Built
with this change and verified that that was no longer happening.
BRANCH=None

Change-Id: Id25d941b988b372e27df979a8a1bbdc331bab900
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/32348
Reviewed-by: Ronald G. Minnich <rminnich@google.com>