depthcharge: Copy FmapArea structures out instead of using them in place.
The FMAP data structures are not set up in such a way that their elements are
often aligned properly. When accessing them on an architecture that cares
about that sort of thing (ARM), there's a good chance something bad will
happen. This change adjusts the FMAP API so that instead of returning a
pointer to the FmapArea structures in place in the actual FMAP, it instead
copies them out into an FmapArea provided by the caller. The copy gives us a
chance to fix the alignment, and it only adds a tiny bit of overhead.

BUG=chrome-os-partner:17340
TEST=Built and booted into depthcharge for Daisy and saw hangs from unaligned
accesses. With this change, the hangs went away.
BRANCH=None

Change-Id: I72fd8fae3670dbf3aff0e7b3fa00f8a585cb2c36
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/32517
Reviewed-by: Ronald G. Minnich <rminnich@google.com>
Reviewed-by: David Hendricks <dhendrix@google.com>