netboot: Add infrastructure for plugging in generic USB drivers.
When libpayload tells us about a new device, we go through all the drivers
that have registered themselves to see if they want to control that device.
When libpayload tells us the device went away, we tell the driver about it so
it can clean up.

BUG=chrome-os-partner:16688
TEST=With a later change that adds a driver for a USB dongle, built and booted
into the netboot payload on Link and saw that the device was probed correctly.
BRANCH=None

Change-Id: Ica2c5e2a20cc8e84d4a45b2f0b0fb19ff53964cd
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/32734