netboot: Add a driver for ASIX based USB network dongles.
This driver will attach to dongles that have vendor and product IDs which
exactly match the dongle I had at my desk. It might need to accept more IDs in
the future, but I'll wait and see what's needed.

BUG=chrome-os-partner:16688
TEST=Built and booted into the netboot payload on Link. Saw that the device
was probed successfully, initialized without error, that it returned the
correct MAC address, that it waited for a link correctly, and that the LEDs on
the dongle were active once the driver was loaded.
BRANCH=None

Change-Id: I08a806a01a62944bd0baab13ea2755aeee029431
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/32736