depthcharge: Add an MKBP based nvstorage option for vboot.
One of the ways of storing vboot's non-volatile data is on the ChromeOS EC,
and that happens to be where it's being stored by the standard Snow firmware.
This set of callbacks is mostly just a wrapper around commands provided by the
MKBP driver itself.

BUG=chrome-os-partner:17340
TEST=Built for and booted into depthcharge on Snow. Saw it attempt to talk to
the ChromeOS EC. Because there's no I2C driver yet that failed.
BRANCH=None

Change-Id: Icc0d3a489be827a33bf3918ffbf8e835e28c0dcd
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/32769
Reviewed-by: Ronald G. Minnich <rminnich@google.com>