netboot: Enable a graphical console.
It will be much easier for people using the netbooting payload if they can see
what's going on on the device's display. This change enables the graphical
console, and also hijacks some vboot mechanisms to ensure that the option rom
has been run.

BUG=chrome-os-partner:16688
TEST=Built and booted on Link where the option rom wasn't being automatically
loaded. Saw that the machine rebooted properly, and that the graphical console
showed system log messages.
BRANCH=None

Change-Id: Id9aada1e8a917dbc73ea4d6b9aa6d0a838e0eebd
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/32742
Reviewed-by: Randall Spangler <rspangler@google.com>