depthcharge: Add an I2C driver for use on Snow.
There are some things in this driver that don't really belong, for instance
the addresses of the various bus interfaces, or code for manipulating GPIOs.
Those will be going away in the future, but some things will need to be
restructured to make that possible.

BUG=chrome-os-partner:17340
TEST=Built and booted on depthcharge and saw the EC initialization succeed and
the EC version string reported correctly.
BRANCH=None

Change-Id: I97c63af725be0cc9ddd0825ce8c672ba84ca6dd2
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/32774
Reviewed-by: Ronald G. Minnich <rminnich@google.com>