depthcharge: Split out where the commonparams come from by arch.
On x86 we get the location through an ACPI data structure in the coreboot
tables, but on ARM (historically) this has been passed to the OS using
different means. I lieu of those mechanisms being unified, depthcharge needs
to support both.

BUG=chrome-os-partner:17340
TEST=Built and booted into depthcharge on Daisy. Saw that it got farther than
it used to. Built and booted on Link.
BRANCH=None

Change-Id: I285591cf47cafa792207d8e26d34f38ebdf6a23c
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/32776
Reviewed-by: Ronald G. Minnich <rminnich@google.com>