depthcharge: Add a driver for the Exynos MCT (multi-core timer).
This is a good system level timer which can back functions like udelay and
provide timer values for timestamps and vboot.

BUG=chrome-os-partner:17340
BUG=chrome-os-partner:17334
TEST=Built and booted into depthcharge on Snow. Saw some timing related I2C
errors go away. Put in an artificial 30s delay and timed it against a stop
watch. Verified that the delay was 30s, as far as I could tell by watching it.
BRANCH=None

Change-Id: I33f9593039057cee8e9ccff57469d0c2210877de
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/32859
Reviewed-by: Ronald G. Minnich <rminnich@google.com>