depthcharge: Generalize and redistribute timekeeping code.
The timekeeping code in libpayload was dependent on rdtsc, and when it was
split up by arch, that code was duplicated even though it was mostly the same.
This change factors out actually reading the count from the timer and the
speed of the timer and puts the definitions of ndelay, udelay, mdelay and
delay into generic code. Then, in x86, the timer_hz and timer_get_raw_value
functions which used to be in depthcharge were moved over to libpayload's
arch/x86/timer.c. In ARM where there isn't a single, canonical timer, those
functions are omitted with the intention that they'll be implemented by a
specific timer driver chosen elsewhere.

BUG=chrome-os-partner:17340
BUG=chrome-os-partner:17334
TEST=Built and booted depthcharge on Link and Daisy.
BRANCH=None

CQ-DEPEND=CL:*32858

Change-Id: If1b18658deb9fffd38c70920e2b16f8e184e410d
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/32894
Reviewed-by: Stefan Reinauer <reinauer@google.com>